4001 - Move getODPHeader
